requirements.txt
openai
fastapi
pymongo
python-dotenv
python-multipart
dnspython
uvicorn
jinja2
httpx
itsdangerous
authlib